iommu/vt-d: Use unsigned long for pages in intel_flush_iotlb_psi()
This changes the prototype of @pages as most of the code uses such a type.

+++ drivers/iommu/intel/iommu.c
@@ -1370,7 +1370,7 @@ static void domain_flush_pasid_iotlb(struct intel_iommu *iommu,
 
 static void iommu_flush_iotlb_psi(struct intel_iommu *iommu,
 				  struct dmar_domain *domain,
-				  unsigned long pfn, unsigned int pages,
+				  unsigned long pfn, unsigned long int pages,
 				  int ih)
 {
 	unsigned int aligned_pages = __roundup_pow_of_two(pages);
